requirements.txt
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
cycler==0.11.0
Flask==2.0.3
fonttools==4.31.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
Werkzeug==2.0.3
